Raycaster working to wolfenstein spec
There looks to be a visual error at 90 degrees, but can't see it
Note that there's temporary code for a collision-based ray caster in there too, but it's not activated yet. Collision comes from the tiles only.
Some tweaks to setup the map to the correct size, add cursor key input and display single/all rays